♻️ observation datatable service created refactored
