+++ src/main/java/com/strandls/observation/service/Impl/ObservationServiceModule.java
@@ -6,6 +6,7 @@ package com.strandls.observation.service.Impl;
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
 import com.strandls.observation.service.MailService;
+import com.strandls.observation.service.ObservationDataTableService;
 import com.strandls.observation.service.ObservationListService;
 import com.strandls.observation.service.ObservationService;
 import com.strandls.observation.service.RecommendationService;
@@ -22,6 +23,7 @@ public class ObservationServiceModule extends AbstractModule {
 		bind(RecommendationService.class).to(RecommendationServiceImpl.class).in(Scopes.SINGLETON);
 		bind(ObservationListService.class).to(ObservationListServiceImpl.class).in(Scopes.SINGLETON);
 		bind(MailService.class).to(MailServiceImpl.class).in(Scopes.SINGLETON);
+		bind(ObservationDataTableService.class).to(ObservationDataTableServiceImpl.class).in(Scopes.SINGLETON);
 		bind(ObservationMapperHelper.class).in(Scopes.SINGLETON);
 		bind(LogActivities.class).in(Scopes.SINGLETON);
 		bind(UserGroupPostingFilterThread.class).in(Scopes.SINGLETON);
